tratando erro 404 e finalizando o curso

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import br.com.alura.forum.controller.Form.AtualizacaoTopicoForm;
 import br.com.alura.forum.controller.dto.DetalhesTopicoDto;
@@ -42,6 +43,7 @@ public class TopicosController {
 	}
 	
 	@PostMapping
+	@Transactional
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico = form.converter(cursoRepository);
 		topicoRepository.save(topico);
@@ -60,23 +62,38 @@ public class TopicosController {
 	}*/
 
 	@GetMapping("/{id}")
-	public DetalhesTopicoDto detalhar(@PathVariable long id){
-		Topico topico = topicoRepository.getById(id);
-		return new DetalhesTopicoDto(topico);
+	public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable long id){
+		Optional<Topico> topico = topicoRepository.findById(id);
+
+		if(topico.isPresent()){
+			return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
+		}
+		return ResponseEntity.notFound().build();
 	}
 
-	@PutMapping("{/id}")
+	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-		Topico topico = form.atualizar(id,topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+
+		Optional<Topico> optional = topicoRepository.findById(id);
+
+		if(optional.isPresent()){
+			Topico topico = form.atualizar(id,topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 
-	@DeleteMapping("{/id}")
-	public ResponseEntity<TopicoDto> deletar(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+	@DeleteMapping("/{id}")
+	@Transactional
+	public ResponseEntity<TopicoDto> deletar(@PathVariable Long id) {
+
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
-
-
 }
